refactor(spec_dependencies): migrate to platform adapter
Closes #299
